gdd: allow private structs to be externally accessed
gdddetector.c contains the core GDD algorithm, we used to put the
structs in it to hide the implementation details, the purpose was to
simplify the detector's usage by providing a small set of carefully
designed APIs.

However some non-algorithm-related helpers, such as the dump functions,
have to be put into this source file to access the struct internals,
which increase the effort to understand and maintain the algorithm.

So now we move the structs to gdddetectorpriv.h, the name still suggests
that it is for private use, however when necessary the callers could
access the struct internals by including this header.
